+++ src/mib_grammar_elixir.erl
@@ -78,9 +78,6 @@ make_module_identity(Name, LU, Org, CI, Desc, Revs, NA) ->
 make_revision(Rev, Desc) ->
     {mc_revision, Rev, Desc}.
 
-make_object_type(Name, Syntax, MaxAcc, Status, Desc, Ref, Kind, NA) ->
-    {mc_object_type, Name, Syntax, undefined, MaxAcc, Status, Desc, Ref, Kind, NA}.
-
 make_object_type(Name, Syntax, Units, MaxAcc, Status, Desc, Ref, Kind, NA) ->
     {mc_object_type, Name, Syntax, Units, MaxAcc, Status, Desc, Ref, Kind, NA}.
 
@@ -485,7 +482,7 @@ yecctoken2string1(Other) ->
 
 
 
--file("src/mib_grammar_elixir.erl", 488).
+-file("src/mib_grammar_elixir.erl", 485).
 
 -dialyzer({nowarn_function, yeccpars2/7}).
 -compile({nowarn_unused_function,  yeccpars2/7}).
@@ -15275,4 +15272,4 @@ yeccpars2_697_(__Stack0) ->
   end | __Stack].
 
 
--file("src/mib_grammar_elixir.yrl", 1219).
+-file("src/mib_grammar_elixir.yrl", 1216).